--- a/Cargo.toml
+++ b/Cargo.toml
@@ -17,6 +17,7 @@ path = "src/main.rs"
 [dependencies]
 chrono = "0.4.38"
 diff = "0.1.13"
+itoa = "1.0.11"
 regex = "1.10.4"
 same-file = "1.0.6"
 unicode-width = "0.2.0"