--- a/tests/integration.rs
+++ b/tests/integration.rs
@@ -15,6 +15,7 @@ use tempfile::{tempdir, NamedTempFile};
 #[test]
 fn unknown_param() -> Result<(), Box<dyn std::error::Error>> {
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg("--foobar");
     cmd.assert()
         .code(predicate::eq(2))
@@ -37,6 +38,7 @@ fn cannot_read_files() -> Result<(), Box<dyn std::error::Error>> {
     let error_message = "The system cannot find the file specified.";
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg(&nopath).arg(file.path());
     cmd.assert()
         .code(predicate::eq(2))
@@ -47,6 +49,7 @@ fn cannot_read_files() -> Result<(), Box<dyn std::error::Error>> {
         )));
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg(file.path()).arg(&nopath);
     cmd.assert()
         .code(predicate::eq(2))
@@ -57,6 +60,7 @@ fn cannot_read_files() -> Result<(), Box<dyn std::error::Error>> {
         )));
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg(&nopath).arg(&nopath);
     cmd.assert().code(predicate::eq(2)).failure().stderr(
         predicate::str::contains(format!(
@@ -74,6 +78,7 @@ fn no_differences() -> Result<(), Box<dyn std::error::Error>> {
     let file = NamedTempFile::new()?;
     for option in ["", "-u", "-c", "-e"] {
         let mut cmd = Command::cargo_bin("diffutils")?;
+        cmd.arg("diff");
         if !option.is_empty() {
             cmd.arg(option);
         }
@@ -93,6 +98,7 @@ fn no_differences_report_identical_files() -> Result<(), Box<dyn std::error::Err
     file1.write_all("foo\n".as_bytes())?;
     for option in ["", "-u", "-c", "-e"] {
         let mut cmd = Command::cargo_bin("diffutils")?;
+        cmd.arg("diff");
         if !option.is_empty() {
             cmd.arg(option);
         }
@@ -111,6 +117,7 @@ fn no_differences_report_identical_files() -> Result<(), Box<dyn std::error::Err
     file2.write_all("foo\n".as_bytes())?;
     for option in ["", "-u", "-c", "-e"] {
         let mut cmd = Command::cargo_bin("diffutils")?;
+        cmd.arg("diff");
         if !option.is_empty() {
             cmd.arg(option);
         }
@@ -135,6 +142,7 @@ fn differences() -> Result<(), Box<dyn std::error::Error>> {
     file2.write_all("bar\n".as_bytes())?;
     for option in ["", "-u", "-c", "-e"] {
         let mut cmd = Command::cargo_bin("diffutils")?;
+        cmd.arg("diff");
         if !option.is_empty() {
             cmd.arg(option);
         }
@@ -155,6 +163,7 @@ fn differences_brief() -> Result<(), Box<dyn std::error::Error>> {
     file2.write_all("bar\n".as_bytes())?;
     for option in ["", "-u", "-c", "-e"] {
         let mut cmd = Command::cargo_bin("diffutils")?;
+        cmd.arg("diff");
         if !option.is_empty() {
             cmd.arg(option);
         }
@@ -178,6 +187,7 @@ fn missing_newline() -> Result<(), Box<dyn std::error::Error>> {
     let mut file2 = NamedTempFile::new()?;
     file2.write_all("bar".as_bytes())?;
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg("-e").arg(file1.path()).arg(file2.path());
     cmd.assert()
         .code(predicate::eq(2))
@@ -194,6 +204,7 @@ fn read_from_stdin() -> Result<(), Box<dyn std::error::Error>> {
     file2.write_all("bar\n".as_bytes())?;
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg("-u")
         .arg(file1.path())
         .arg("-")
@@ -210,6 +221,7 @@ fn read_from_stdin() -> Result<(), Box<dyn std::error::Error>> {
     );
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg("-u")
         .arg("-")
         .arg(file2.path())
@@ -226,6 +238,7 @@ fn read_from_stdin() -> Result<(), Box<dyn std::error::Error>> {
     );
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg("-u").arg("-").arg("-");
     cmd.assert()
         .code(predicate::eq(0))
@@ -235,6 +248,7 @@ fn read_from_stdin() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(unix)]
     {
         let mut cmd = Command::cargo_bin("diffutils")?;
+        cmd.arg("diff");
         cmd.arg("-u")
             .arg(file1.path())
             .arg("/dev/stdin")
@@ -270,6 +284,7 @@ fn compare_file_to_directory() -> Result<(), Box<dyn std::error::Error>> {
     da.write_all(b"da\n").unwrap();
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg("-u").arg(&directory).arg(&a_path);
     cmd.assert().code(predicate::eq(1)).failure();
 
@@ -284,6 +299,7 @@ fn compare_file_to_directory() -> Result<(), Box<dyn std::error::Error>> {
     );
 
     let mut cmd = Command::cargo_bin("diffutils")?;
+    cmd.arg("diff");
     cmd.arg("-u").arg(&a_path).arg(&directory);
     cmd.assert().code(predicate::eq(1)).failure();