… Cargo.lock
… fuzz/Cargo.toml
… fuzz/dictionaries/cmp.txt
… src/params.rs
… tests/run-upstream-testsuite.sh