@sylvestre 2024-09-25
could you please run your benchmarks with hyperfine? time is too limited