@oSoMoN 2024-09-26
clippy is reporting 9 trivial errors, could you address them to get the CI results green?