@oSoMoN 2024-09-27
> I checked with strace that the cmp being run was appropriate fwiw. The last error message from GNU cmp is a real wth xD

I ran `LANG=C ./cmp` from the tests directory in a local checkout of the upstream repository, and I'm not seeing the "invalid --ignore-initial value" error that you're observing. I do see it when using cmp built from your branch.

Note that the `cmp` test script prepends `../src` to `$PATH`. Could it be that you had symlinked the wrong executable in there?
I'm seeing a different behaviour when I run that same test script against my system-wide `/usr/bin/cmp` installed from distro packages (using Ubuntu 24.04).

The following upstream commits seem relevant: [4ee8300](https://git.savannah.gnu.org/cgit/diffutils.git/commit?id=4ee8300) and [9c5fcbd](https://git.savannah.gnu.org/cgit/diffutils.git/commit?id=9c5fcbd).