@oSoMoN 2024-09-27
Comparing further the output of the `cmp` test script, it looks like your implementation conforms to the released version of GNU cmp (version 3.10), i.e. it chokes on large numbers passed to `-i`/`--ignore-initial`.

We have two options here:
 * align your implementation with the latest unreleased changes in the GNU diffutils repository
 * change `tests/run-upstream-testsuite.sh` to checkout a released version of the test suite, not the latest HEAD

I think that both approaches are equally interesting and valid (one favours the bleeding edge, the other compatibility with released versions), so feel free to pick whichever you prefer (I can help with updating the test script if you decide to go that way).