@oSoMoN 2024-09-27
> I just need to investigate this one now:
> 
>     * brief-vs-stat-zero-kernel-lies           SKIP
>     * brief-vs-stat-zero-kernel-lies           FAIL

I looked into it, and it turns out it's an issue in the test suite runner, not in your code. I filed #92 to track it.